My computer made a weird noise so I'm making backups
This isn't all ready for upload but here it is:
Major pre-processor updates:
Added "de-snowing", this allows for pixels belonging to one set of masks to be reassigned to other masks if they meet some condition (generally having too many or too few neighbours of a certain mask type, similar to conway's game of life). This is used to both black out the dust in the empty space surrounding the sample, and correctly account for pixels which rightly belong to either plag or melt inclusions, but are above/below the threshold to differentiate them due to apparently random noise.
The de-snowing and other fixes to the de-gridding process have improved the efficacy of that step by >7 orders of magnitude. Therefore I think I shall now declare the infra-tile gradient dead.
Work is ongoing to extend this to the right-hand edge effect, I anticipate substantial improvements once I have time to solve the bugs.
The tiles are now denoised using a 2d butterworth filter instead of the total-variance method. This does appear to sacrifice some edge sharpness, but runs in many tiles/second rather than many seconds/tile, and seems to very effectively kill the high frequency noise.
These combine to produce images that do not need the gradient or bloom parts of the post-processing, and reach significantly better contrasts.
I have yet to attempt the conversion to composition using these images, but will do post-exams.
I'm not sure what the changes made to the stitcher and brightness matcher are, it is likely that they are the changes I made while trying out higher resolution tiles, and do not constitute major rewrites.